These are changes for v1.7.2.7.  A further fix for the int32/int64 overflow issue, this one related to ExtractProgress. This involves changes to Events.cs, ZipEntry.cs , and ZipFile.cs. Also, some updated test code to verify that this works. Also, some additions to the interface to facilitate ZIP64 handling, and some tests for same. Also, an update to the CreateZip and ZipDir samples for style & clarity.
git-svn-id: https://dotnetzip.svn.codeplex.com/svn@29367 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Zip Partial DLL/Events.cs
@@ -419,7 +419,7 @@ namespace Ionic.Zip
         }
 
 
-        internal static ExtractProgressEventArgs ByteUpdate(string archiveName, ZipEntry entry, int bytesWritten, Int64 totalBytes)
+        internal static ExtractProgressEventArgs ByteUpdate(string archiveName, ZipEntry entry, Int64 bytesWritten, Int64 totalBytes)
         {
             var x = new ExtractProgressEventArgs(archiveName, ZipProgressEventType.Extracting_EntryBytesWritten);
             x.ArchiveName = archiveName;
